added admin route to update missing slugs - fixes #4
Usage: While logged in as admin (or with uniqueslugs manage rights) call `https://cockpit-url.tld/uniqueslugs/updateEntriesWithoutSlug` in your browser. All changed entries are returned with id and the slug field of the default language. All slugs in other languages are updated, too. I'm not motivated to write a GUI right now. Pull requests are welcome.